Removed no more used components
