+++ components/PageHeading.js
@@ -1,4 +1,3 @@
-/* This example requires Tailwind CSS v2.0+ */
 import { ChevronLeftIcon, ChevronRightIcon } from "@heroicons/react/solid";
 
 export const PageHeading = ({ title, breadcumbs, actions }) => {
